requirements.txt
django-environ
Django
# mysqlclient
djangorestframework-simplejwt
djangorestframework
gunicorn
celery
redis
django-filter
psycopg2-binary
twilio
pillow